third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-column-block-start-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 1
  assert_equals: \n<item data-expected-margin-top="0" data-offset-y="8"></item>\noffsetTop expected 8 but got 18
[FAIL] flexbox > item 2
  assert_equals: \n<item data-expected-margin-top="10" data-offset-y="68"></item>\noffsetTop expected 68 but got 78
Harness: the test ran to completion.
